+++ je_auto_control/__init__.py
@@ -50,7 +50,7 @@ from je_auto_control.wrapper.auto_control_record import record
 from je_auto_control.wrapper.auto_control_record import stop_record
 
 # json
-from je_auto_control.utils.action_executer.action_executor import execute_action
+from je_auto_control.utils.action_executor.action_executor import execute_action
 from je_auto_control.utils.action_file.json_file import read_action_json
 from je_auto_control.utils.action_file.json_file import write_action_json
