--- je_auto_control/utils/action_executer/__init__.py
@@ -1,1 +1,0 @@
-from je_auto_control.utils.action_executer import *
