+++ je_auto_control/wrapper/auto_control_record.py
@@ -1,6 +1,6 @@
 import sys
 
-from je_auto_control.utils.action_executer.action_executor import execute_action
+from je_auto_control.utils.action_executor.action_executor import execute_action
 from je_auto_control.utils.exception.exception_tag import macos_record_error
 from je_auto_control.utils.exception.exceptions import AutoControlException
 from je_auto_control.utils.exception.exceptions import AutoControlJsonActionException
@@ -26,12 +26,12 @@ def stop_record():
             new_list.append([action[0], dict([["keycode", action[1]]])])
         else:
             new_list.append([action[0], dict(zip(["mouse_keycode", "x", "y"], [action[0], action[1], action[2]]))])
-    return execute_action(new_list)
+    return new_list
 
 
 if __name__ == "__main__":
     record()
     from time import sleep
     sleep(5)
-    stop_record()
+    print(stop_record())
     sleep(2)
