+++ test/unit_test/keyboard/keyboard_write_test.py
@@ -32,3 +32,4 @@ try:
 except AutoControlKeyboardException as error:
     print(repr(error), file=sys.stderr)
 
+
